fix tests add detekt add ktlint

+++ library/src/commonMain/kotlin/com/viplearner/api/models/ArenaPerf.kt
@@ -1,12 +1,11 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class ArenaPerf(
     val key: PerfType,
     val name: String,
-    val position: Int,
-    val icon: String? = null
+    val position: Long,
+    val icon: String? = null,
 )
