+++ library/src/commonMain/kotlin/com/viplearner/api/models/ArenaRatingObj.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class ArenaRatingObj(
     val perf: PerfType? = null,
-    val rating: Int
+    val rating: Long,
 )
