+++ library/src/commonMain/kotlin/com/viplearner/api/models/ArenaTournamentPlayer.kt
@@ -1,12 +1,11 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class ArenaTournamentPlayer(
-    val games: Int,
-    val score: Int,
-    val rank: Int,
-    val performance: Int? = null
+    val games: Long,
+    val score: Long,
+    val rank: Long,
+    val performance: Long? = null,
 )
