+++ library/src/commonMain/kotlin/com/viplearner/api/models/BroadcastPlayerEntry.kt
@@ -1,20 +1,19 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class BroadcastPlayerEntry(
     val name: String,
     val title: Title? = null,
-    val rating: Int? = null,
-    val fideId: Int? = null,
+    val rating: Long? = null,
+    val fideId: Long? = null,
     val team: String? = null,
     val fed: String? = null,
     val score: Double? = null,
-    val played: Int? = null,
-    val ratingDiff: Int? = null,
-    val performance: Int? = null,
+    val played: Long? = null,
+    val ratingDiff: Long? = null,
+    val performance: Long? = null,
     val tiebreaks: List<BroadcastPlayerTiebreak>? = null,
-    val rank: Int? = null
+    val rank: Long? = null,
 )
