+++ library/src/commonMain/kotlin/com/viplearner/api/models/BroadcastPlayerEntryWithFideAndGames.kt
@@ -1,18 +1,17 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class BroadcastPlayerEntryWithFideAndGamesFide(
-    val year: Int? = null,
-    val ratings: Ratings? = null
+    val year: Long? = null,
+    val ratings: Ratings? = null,
 ) {
     @Serializable
     data class Ratings(
-        val standard: Int? = null,
-        val rapid: Int? = null,
-        val blitz: Int? = null
+        val standard: Long? = null,
+        val rapid: Long? = null,
+        val blitz: Long? = null,
     )
 }
 
@@ -22,14 +21,14 @@ data class BroadcastPlayerEntryWithFideAndGames(
     val games: List<BroadcastGameEntry>? = null,
     val name: String,
     val title: Title? = null,
-    val rating: Int? = null,
-    val fideId: Int? = null,
+    val rating: Long? = null,
+    val fideId: Long? = null,
     val team: String? = null,
     val fed: String? = null,
     val score: Double? = null,
-    val played: Int? = null,
-    val ratingDiff: Int? = null,
-    val performance: Int? = null,
+    val played: Long? = null,
+    val ratingDiff: Long? = null,
+    val performance: Long? = null,
     val tiebreaks: List<BroadcastPlayerTiebreak>? = null,
-    val rank: Int? = null
+    val rank: Long? = null,
 )
