+++ library/src/commonMain/kotlin/com/viplearner/api/models/Clock.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class Clock(
-    val limit: Int,
-    val increment: Int
+    val limit: Long,
+    val increment: Long,
 )
