+++ library/src/commonMain/kotlin/com/viplearner/api/models/CloudEval.kt
@@ -1,16 +1,16 @@
 package com.viplearner.api.models
 
-import kotlinx.serialization.Serializable
 import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonContentPolymorphicSerializer
 import kotlinx.serialization.json.jsonObject
 
 @Serializable
 data class CloudEval(
-    val depth: Int,
+    val depth: Long,
     val fen: String,
-    val knodes: Int,
-    val pvs: List<Pv>
+    val knodes: Long,
+    val pvs: List<Pv>,
 ) {
     class PvSerializer : JsonContentPolymorphicSerializer<Pv>(Pv::class) {
         override fun selectDeserializer(element: kotlinx.serialization.json.JsonElement) =
@@ -26,15 +26,15 @@ data class CloudEval(
         @Serializable
         @SerialName("Non-mate variation")
         data class NonMateVariation(
-            val cp: Int,
-            val moves: String
+            val cp: Long,
+            val moves: String,
         ) : Pv()
 
         @Serializable
         @SerialName("Mate variation")
         data class MateVariation(
-            val mate: Int,
-            val moves: String
+            val mate: Long,
+            val moves: String,
         ) : Pv()
     }
 }
