+++ library/src/commonMain/kotlin/com/viplearner/api/models/Count.kt
@@ -1,21 +1,20 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class Count(
-    val all: Int,
-    val rated: Int,
-    val ai: Int? = null,
-    val draw: Int,
-    val drawH: Int? = null,
-    val loss: Int,
-    val lossH: Int? = null,
-    val win: Int,
-    val winH: Int? = null,
-    val bookmark: Int,
-    val playing: Int,
-    val import: Int,
-    val me: Int
+    val all: Long,
+    val rated: Long,
+    val ai: Long? = null,
+    val draw: Long,
+    val drawH: Long? = null,
+    val loss: Long,
+    val lossH: Long? = null,
+    val win: Long,
+    val winH: Long? = null,
+    val bookmark: Long,
+    val playing: Long,
+    val import: Long,
+    val me: Long,
 )
