+++ library/src/commonMain/kotlin/com/viplearner/api/models/Crosstable.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class Crosstable(
-    val users: Map<String, Int>,
-    val nbGames: Int
+    val users: Map<String, Double>,
+    val nbGames: Long,
 )
