+++ library/src/commonMain/kotlin/com/viplearner/api/models/GameEventPlayer.kt
@@ -1,14 +1,13 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class GameEventPlayer(
-    val aiLevel: Int? = null,
+    val aiLevel: Long? = null,
     val id: String,
     val name: String,
     val title: Title? = null,
-    val rating: Int? = null,
-    val provisional: Boolean? = null
+    val rating: Long? = null,
+    val provisional: Boolean? = null,
 )
