+++ library/src/commonMain/kotlin/com/viplearner/api/models/GameOpening.kt
@@ -1,11 +1,10 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class GameOpening(
     val eco: String,
     val name: String,
-    val ply: Int
+    val ply: Long,
 )
