+++ library/src/commonMain/kotlin/com/viplearner/api/models/OpeningExplorerGamePlayer.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class OpeningExplorerGamePlayer(
     val name: String,
-    val rating: Int
+    val rating: Long,
 )
