+++ library/src/commonMain/kotlin/com/viplearner/api/models/OpponentGoneEvent.kt
@@ -1,11 +1,10 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class OpponentGoneEvent(
     val type: String,
     val gone: Boolean,
-    val claimWinInSeconds: Int? = null
+    val claimWinInSeconds: Long? = null,
 )
