+++ library/src/commonMain/kotlin/com/viplearner/api/models/Perf.kt
@@ -1,13 +1,12 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class Perf(
-    val games: Int,
-    val rating: Int,
-    val rd: Int,
-    val prog: Int,
-    val prov: Boolean? = null
+    val games: Long,
+    val rating: Long,
+    val rd: Long,
+    val prog: Long,
+    val prov: Boolean? = null,
 )
