+++ library/src/commonMain/kotlin/com/viplearner/api/models/PlayTime.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class PlayTime(
-    val total: Int,
-    val tv: Int
+    val total: Long,
+    val tv: Long,
 )
