+++ library/src/commonMain/kotlin/com/viplearner/api/models/Profile.kt
@@ -1,7 +1,6 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class Profile(
@@ -9,11 +8,11 @@ data class Profile(
     val location: String? = null,
     val bio: String? = null,
     val realName: String? = null,
-    val fideRating: Int? = null,
-    val uscfRating: Int? = null,
-    val ecfRating: Int? = null,
-    val cfcRating: Int? = null,
-    val rcfRating: Int? = null,
-    val dsbRating: Int? = null,
-    val links: String? = null
+    val fideRating: Long? = null,
+    val uscfRating: Long? = null,
+    val ecfRating: Long? = null,
+    val cfcRating: Long? = null,
+    val rcfRating: Long? = null,
+    val dsbRating: Long? = null,
+    val links: String? = null,
 )
