+++ library/src/commonMain/kotlin/com/viplearner/api/models/PuzzleModePerf.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class PuzzleModePerf(
-    val runs: Int,
-    val score: Int
+    val runs: Long,
+    val score: Long,
 )
