+++ library/src/commonMain/kotlin/com/viplearner/api/models/PuzzlePerformance.kt
@@ -1,13 +1,12 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class PuzzlePerformance(
-    val firstWins: Int,
-    val nb: Int,
-    val performance: Int,
-    val puzzleRatingAvg: Int,
-    val replayWins: Int
+    val firstWins: Long,
+    val nb: Long,
+    val performance: Long,
+    val puzzleRatingAvg: Long,
+    val replayWins: Long,
 )
