+++ library/src/commonMain/kotlin/com/viplearner/api/models/RatingHistoryEntry.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class RatingHistoryEntry(
     val name: String? = null,
-    val points: List<List<Int>>? = null
+    val points: List<List<Long>>? = null,
 )
