+++ library/src/commonMain/kotlin/com/viplearner/api/models/TeamPaginatorJson.kt
@@ -1,15 +1,14 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class TeamPaginatorJson(
-    val currentPage: Int,
-    val maxPerPage: Int,
+    val currentPage: Long,
+    val maxPerPage: Long,
     val currentPageResults: List<Team>,
-    val previousPage: Int?,
-    val nextPage: Int?,
-    val nbResults: Int,
-    val nbPages: Int
+    val previousPage: Long?,
+    val nextPage: Long?,
+    val nbResults: Long,
+    val nbPages: Long,
 )
