+++ library/src/commonMain/kotlin/com/viplearner/api/models/TeamRequest.kt
@@ -1,12 +1,11 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class TeamRequest(
     val teamId: String,
     val userId: String,
-    val date: Int,
-    val message: String? = null
+    val date: Long,
+    val message: String? = null,
 )
