+++ library/src/commonMain/kotlin/com/viplearner/api/models/UserActivityFollowList.kt
@@ -1,10 +1,9 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class UserActivityFollowList(
     val ids: List<String>,
-    val nb: Int? = null
+    val nb: Long? = null,
 )
