+++ library/src/commonMain/kotlin/com/viplearner/api/models/UserActivityScore.kt
@@ -1,18 +1,17 @@
 package com.viplearner.api.models
 
 import kotlinx.serialization.Serializable
-import kotlinx.serialization.SerialName
 
 @Serializable
 data class UserActivityScoreRp(
-    val before: Int,
-    val after: Int
+    val before: Long,
+    val after: Long,
 )
 
 @Serializable
 data class UserActivityScore(
-    val win: Int,
-    val loss: Int,
-    val draw: Int,
-    val rp: UserActivityScoreRp
+    val win: Long,
+    val loss: Long,
+    val draw: Long,
+    val rp: UserActivityScoreRp,
 )
